releases/scripts/requirements.txt
pygerrit2 < 2.1.0
PyYAML < 4.0
jsonschema < 2.7.0
jenkins-job-builder
